requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
sqlparse==0.5.3
typing_extensions==4.15.0